database_requirements.txt
google-cloud-datastore<2.0.0dev
google-cloud-ndb